Source Receptor Matrix Tools

InMAPSRM.connect_to_isrm! — Function

connect_to_isrm!(; region="us-east-2", creds = nothing, url="s3://inmap-model/isrm_v1.2.1.zarr/")

Connects to AWS, opens the INMAP source receptor matrix with zopen (from Zarr.jl)

InMAPSRM.run_sr — Function

run_sr(source_emis::DataFrame, [cell_variables]) -> receptor_emis::DataFrame

Runs the source receptor matrix for the given source emissions, where each row represents an emitter. Note that source_emis must have the following columns:

latitude <: Number - latitude of the source

longitude <: Number - longitude of the source

layer_idx <: Integer - layer index ∈ {1,2,3}, for ground-level, low-stack, and high-stack emissions, respectively

<emis_type> <: Number - the average emissions rate, in μg / s, for emission type <emis_type> ∈ {PM2_5, NOx, SO2, VOC, NH3}. There can be between 1 and 5 of these columns.

(optional) source_idx <: Int64 - which grid cell index corresponds to the source. If given, latitude and longitude no longer used.

Returns the receptor_emis table which contains columns for each of the PM types created by the source emissions, as well as columns for each of the variables in cell_variables, and a column for geometry_longlat, which contains the (lon, lat) geometry of each grid cell.

InMAPSRM.SRM — Function

SRM(emis_type) -> srm::Array{Float32, 3}

Returns source receptor matrix retrieved from fs, with emis_type ∈ ("PrimaryPM25", "SOA", "pNO3", "pSO4", "pNH4")

Note that these are huge (52411 x 52411 x 3) and will require ~60-70GB of RAM to hold onto. Index via (receptor, source, layer). Care should be taken if multiple stored in memory. May take 5-10 minutes to fetch from AWS.

To create a sparse SRM for only a select group of source locations, see make_sparse_srm. (that will still call SRM, but will call garbage collection before returning the sparse matrix)

InMAPSRM.SparseSRM — Type

struct SparseSRM <: AbstractArray{Float32, 3}

A simple type used for storing a sparse array representation of an SRM.

InMAPSRM.compute_receptor_emis — Function

compute_receptor_emis(srm, source_idx(s), layer_idx(s), val(s)) -> receptor_emis::Vector

Compute the emissions at each receptor from srm for each source specified by source_idx and layer_idx, which correspond to val(s)

compute_receptor_emis(srm, source_emis::Matrix) -> receptor_emis::Vector

Compute the emissions at each receptor from srm for source_emis, a NSR x 3 matrix containing annual average emission rates at each grid cell and layer, in units of micrograms per second.

srm can be any of the following types:

Array{Float64, 3} - returned by SRM

SparseSRM

String - pm emission type ∈ ("PrimaryPM25", "SOA", "pNO3", "pSO4", "pNH4")

Cell Data

InMAPSRM.get_isrm_variable_summary — Function

get_isrm_variable_summary() -> summary::DataFrame

Returns a summary of the variables in the source receptor matrix

InMAPSRM.get_isrm_cell_data — Function

get_isrm_cell_data(vars::AbstractVector{<:AbstractString}; geometry=true) -> cell_data::DataFrame

To see list of available variables, see get_isrm_variable_summary

get_isrm_cell_data(var::AbstractString) -> v::Vector

Fetch the grid-cell data for the variable var.

InMAPSRM.get_isrm_fs — Function

get_isrm_fs() -> fs

Return the AWS file system.

Geometry

Missing docstring.

Missing docstring for longlat2irsm. Check Documenter's build log for details.

InMAPSRM.isrm2longlat — Function

isrm2longlat() -> trans

Returns a Proj.Transformation for converting to (lon, lat) from the LCC coordinate system used by ISRM:

+proj=lcc +lat_1=33.000000 +lat_2=45.000000 +lat_0=40.000000 +lon_0=-97.000000 +x_0=0 +y_0=0 +a=6370997.000000 +b=6370997.000000 +to_meter=1.

InMAPSRM.get_isrm_cell_geom — Function

get_isrm_cell_geom() -> geom::Vector{Box{2, Float64}}

InMAPSRM.get_isrm_cell_geom_longlat — Function

get_isrm_cell_geom_longlat(geom::Vector{Box{2, Float64}}) -> geom::Vector{Quadrangle{2, Float64}}

Returns the cell geometry of the SRM in longlat form.

InMAPSRM.get_cell_idxs — Function

get_cell_idxs(longlats, cell_geom::Vector; threshold = 1000) -> cell_idxs

get_cell_idxs(longs, lats, cell_geom::Vector; threshold = 1000) -> cell_idxs

Return the indices of the closest cells (in cell_geom) for each of the longs and lats supplied. If there is no grid cell within threshold meters from the point in geo_df, gives an index of 0. cell_geom can either be the original geometry (represented as Boxes), or the longlat geometry (represented as Quadrangles).